### 2021-12-29 #4026 @whedon
[PRE REVIEW]: SimuPy Flight Vehicle Toolkit
**Submitting author:** @ixjlyons (<a href="http://orcid.org/0000-0002-9143-8459">Kenneth Lyons</a>)
**Repository:** https://github.com/nasa/simupy-flight
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
