### 2021-12-29 #4026 @arfon
@ixjlyons – many thanks for your submission to JOSS. While we look for an editor to handle this submission, if you want to fix the DOIs (noting that Whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
