### 2021-12-29 #4026 @arfon
@ixjlyons – I also see that your software is licensed under the NOSA agreement. Could you please update the PDF license to be a plain text file named `LICENSE` in your repository (with the [NOSA agreement language](https://opensource.org/licenses/NASA-1.3)).
