+++ paper/references.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 @article{simupy,
   title={{SimuPy}: A {Python} framework for modeling and simulating dynamical systems},
   author={Margolis, Benjamin WL},
@@ -37,21 +38,19 @@
   year={2015}
 }
 
-@article{sympy,
- title = {SymPy: symbolic computing in Python},
- author = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k, Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K. and Singh, Sartaj and Rathnayake, Thilina and Vig, Sean and Granger, Brian E. and Muller, Richard P. and Bonazzi, Francesco and Gupta, Harsh and Vats, Shivam and Johansson, Fredrik and Pedregosa, Fabian and Curry, Matthew J. and Terrel, Andy R. and Rou\v{c}ka, \v{S}t\v{e}p\'{a}n and Saboo, Ashutosh and Fernando, Isuru and Kulal, Sumith and Cimrman, Robert and Scopatz, Anthony},
- year = 2017,
- month = jan,
- keywords = {Python, Computer algebra system, Symbolics},
- abstract = {
-            SymPy is an open source computer algebra system written in pure Python. It is built with a focus on extensibility and ease of use, through both interactive and programmatic applications. These characteristics have led SymPy to become a popular symbolic library for the scientific Python ecosystem. This paper presents the architecture of SymPy, a description of its features, and a discussion of select submodules. The supplementary material provide additional examples and further outline details of the architecture and features of SymPy.
-         },
- volume = 3,
- pages = {e103},
- journal = {PeerJ Computer Science},
- issn = {2376-5992},
- url = {https://doi.org/10.7717/peerj-cs.103},
- doi = {10.7717/peerj-cs.103}
+@Article{sympy,
+  author   = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k, Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K. and Singh, Sartaj and Rathnayake, Thilina and Vig, Sean and Granger, Brian E. and Muller, Richard P. and Bonazzi, Francesco and Gupta, Harsh and Vats, Shivam and Johansson, Fredrik and Pedregosa, Fabian and Curry, Matthew J. and Terrel, Andy R. and Rou\v{c}ka, \v{S}t\v{e}p\'{a}n and Saboo, Ashutosh and Fernando, Isuru and Kulal, Sumith and Cimrman, Robert and Scopatz, Anthony},
+  title    = {SymPy: symbolic computing in Python},
+  doi      = {10.7717/peerj-cs.103},
+  issn     = {2376-5992},
+  pages    = {e103},
+  url      = {https://doi.org/10.7717/peerj-cs.103},
+  volume   = {3},
+  abstract = {SymPy is an open source computer algebra system written in pure Python. It is built with a focus on extensibility and ease of use, through both interactive and programmatic applications. These characteristics have led SymPy to become a popular symbolic library for the scientific Python ecosystem. This paper presents the architecture of SymPy, a description of its features, and a discussion of select submodules. The supplementary material provide additional examples and further outline details of the architecture and features of SymPy.},
[…]
