+++ paper/references.bib
@@ -116,7 +116,7 @@
   author    = {Margolis, Benjamin and Okolo, Wendy and D'Souza, Sarah N and Johnson, Breanna J},
   booktitle = {AIAA Scitech 2021 Forum},
   title     = {Pterodactyl: Guidance and Control of a Symmetric Deployable Entry Vehicle using an Aerodynamic Control System},
-  doi       = {10.2514/6.2021-0764.},
+  doi       = {10.2514/6.2021-0764},
   pages     = {0764},
   year      = {2021},
 }
