### 2022-01-13 #4026 @prashjha
@ixjlyons, while you and I are looking for potential reviewers, I have a few remarks for the draft of the paper:

As per the example JOSS paper [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), you are missing 'Statement of Need'. For a software article, I would also like to see some details about the model and application using your code. You may also need to add the 'Acknowledgments' section if any. I would recommend keeping the article within 3 page limit (excluding references).
