### 2022-02-22 #4026 @prashjha
Hi @aliaksei135, would you be interested in reviewing this submission for JOSS?
