### 2022-02-24 06de4eb
first draft of docstrings
