+++ simupy_flight/__init__.py
@@ -14,27 +14,112 @@ earth_mean_radius = 6371007.1809
 
 # kinematic functions
 def get_nonflat_pc2pd(a, f):
+    """
+    Generate a planetocentric to planetodetic coordinate transformation function for an
+    ellipsoidal planetary model, parameterized by equitorial radius `a` and flattening `f`.
+    The transformation function takes in the planetocentric
+    rectangular coordinates `px`, `py`, `pz` and returns the planetodetic `longitude`, `latitude`,
+    and `altitude`. The units of `a`, `px`, `py`, `pz`, and `altitude` are all the same and
+    assumed to be meters to be consistent with other kinematics.
+    Uses the ERFA implementation of Fukushima's method.
+
+    Parameters
+    ----------
+    a : float
+        equitorial radius
+    f : float
+        flattening parameter
+    
+    Returns
+    -------
+    pcf2pd : callable
+        The planetocentric to planetodetic coordinate trnasformation function.
+    """
     def pcf2pd(px, py, pz):
         return np.array(erfa.gc2gde(a, f, np.array([px, py, pz])))
     return pcf2pd
 
 def get_flat_pc2pd():
+    """
+    Generate a planetocentric to planetodetic coordinate transformation function for a
+    flat planetary model.
+    The transformation function takes in the planetocentric
+    rectangular coordinates `px`, `py`, `pz` and returns a constant 0 for `longitude`, 
+    -pi/2 for `latitutde`, and uses `pz` for `altitude`.
+    
+    Returns
+    -------
+    pcf2pd : callable
+        The planetocentric to planetodetic coordinate trnasformation function.
+    """
     def pcf2pd(px, py, pz):
         return np.array([0.0, -np.pi/2, pz])
     return pcf2pd
 
 def get_nonflat_pd2pc(a, f):
+    """
+    Generate a planetodetic to planetocentric coordinate transformation function for an
+    ellipsoidal planetary model, parameterized by equitorial radius`$a` and flattening `f`.
+    The transformation function takes in the position defined by the planetodetic 
+    `longitude`, `latitude`, and `altitude` and returns the planetocentric
+    rectangular coordinates `px`, `py`, `pz`.
+    The units of `a`, `px`, `py`, `pz`, and `altitude` are all the same and
+    assumed to be meters to be consistent with other kinematics.
+    Uses the ERFA implementation of Fukushima's method.
+
+    Parameters
+    ----------
+    a : float
+        equitorial radius
+    f : float
+        flattening parameter
+    
+    Returns
+    -------
+    pcf2pd : callable
+        The planetodetic to planetocentric coordinate trnasformation function.
+    """
     def pd2pcf(longitude, latitude, altitude):
         return np.array(erfa.gd2gce(a, f, longitude, latitude, altitude))
     return pd2pcf
 
 def get_flat_pd2pc():
+    """
+    Generate a planetodetic to planetocentric coordinate transformation function for a
+    flat planetary model.
+    The transformation function takes in the planetocentric
+    rectangular coordinates px, py, pz and returns a constant 0 longitude, 
+    -90 degree latitutde, and uses pz for altitude.
+    
+    Returns
+    -------
+    pcf2pd : callable
+        The planetocentric to planetodetic coordinate trnasformation function.
+    """
     def pd2pcf(longitude, latitude, altitude):
         return np.array([longitude, latitude, altitude])
     return pd2pcf
 
 class Planetodetic:
+    """
+    Namespace class used to model the planetodetics used by the Planet class.
+    """
+
     def __init__(self, a=0., f=0., omega_p=0.):
+        """
+        Parameters
+        ----------
+        a : float
+            equitorial radius
+        f : float
[…]
