### 2022-02-24 f651343
adding helper functions with docstrings
