### 2022-02-25 e365fbf
Fix up a few typos, add instructions for getting test deps
Also change extras name from nesc to test (more common) and fix license
field in setup.py.
