### 2022-02-25 0578714
starting to refactor case 11
