+++ nesc_test_cases/F16_model.py
@@ -1,47 +1,66 @@
-# run after process_NESC_DaveML
+"""
+===========
+F16 Model
+===========
+
+
+
+"""
 import F16_aero
 import F16_prop
 import F16_inertia
 import numpy as np
 import simupy_flight
-from nesc_testcase_helper import ft_per_m, kg_per_slug, N_per_lbf
+from nesc_testcase_helper import ft_per_m, kg_per_slug
+
+S_A = 300.0/(ft_per_m**2)
+b_l = 30./ft_per_m
+c_l = 11.32/ft_per_m
+a_l = b_l
+
+N_per_lbf = 4.44822
+
+Ixx = 9_496.0*kg_per_slug/(ft_per_m**2) #slug-ft2
+Iyy = 55_814.0*kg_per_slug/(ft_per_m**2) #slug-ft2
+Izz = 63_100.0*kg_per_slug/(ft_per_m**2) #slug-ft2
+Ixy = 0.0*kg_per_slug/(ft_per_m**2) #slug-ft2
+Iyz = 0.0*kg_per_slug/(ft_per_m**2) #slug-ft2
+Izx = 982.0*kg_per_slug/(ft_per_m**2) #slug-ft2
+m = 637.26*kg_per_slug #slug
+
+longitudinal_com_pos_percent = 25.
+x_com = 0.01*c_l*(35.-longitudinal_com_pos_percent)
+y_com = 0.
+z_com = 0.
 
 x_mrc = 0.
 y_mrc = 0.
 z_mrc = 0.
 
+# Ixx, Iyy, Izz, Ixy, Iyz, Izx, m, x_com, y_com, z_com = F16_inertia.F16_inertia(0.25)
 inertia_output = F16_inertia.F16_inertia(25)
 Ixx, Iyy, Izz, Izx, Ixy, Iyz = inertia_output[:6]*kg_per_slug/(ft_per_m**2)
-m = 637.26*kg_per_slug #slugs; DaveML parsed data does not match documentation
-x_com, y_com, z_com = inertia_output[[-1, -3, -2]]/ft_per_m
 
-aero_output = F16_aero.F16_aero(*np.zeros(9))
-S_A = aero_output[2]/(ft_per_m**2)
-a_l, b_l, c_l = aero_output[[1, 1, 0]]/ft_per_m
+class F16(simupy_flight.Vehicle):
+    def __init__(self):
+        super().__init__(
+            m=m,
+            I_xx=Ixx, I_yy=Iyy, I_zz=Izz,
+            I_xy=Ixy, I_yz=Iyz, I_xz=Izx,
+            x_com=x_com, y_com=y_com, z_com=z_com,
 
+            base_aero_coeffs=0.,
+            x_mrc=x_mrc, y_mrc=y_mrc, z_mrc=z_mrc,
+            S_A=S_A, a_l=a_l, b_l=b_l, c_l=c_l, d_l=0.,
 
-def F16_prop_model(*args):
-    """
-    Wrapper to convert units appropriately
-    """
-    throttle = args[-1]
-    alt = args[simupy_flight.Vehicle.h_D_arg_idx]*ft_per_m
-    Ma = args[simupy_flight.Vehicle.Ma_arg_idx]
-    return F16_prop.F16_prop(throttle, alt, Ma)*N_per_lbf
-
-class VehicleWithAlternateAeroSturcture(simupy_flight.Vehicle):
-    def tot_aero_forces_moments(self, qbar, Ma, Re, V_T, alpha, beta, p_B, q_B, r_B, el, ail, rdr, *args):
-        """
-        Alternate aero structure that depends only on base_aero_coeffs with three inputs
-        (elevator, aileron, and rudder), forces in body-fixed coordinates instead of wind-fixed,
[…]
