+++ simupy_flight/__init__.py
@@ -16,7 +16,7 @@ earth_mean_radius = 6371007.1809
 def get_nonflat_pc2pd(a, f):
     """
     Generate a planetocentric to planetodetic coordinate transformation function for an
-    ellipsoidal planetary model, parameterized by equitorial radius `a` and flattening `f`.
+    ellipsoidal planetary model, parameterized by equatorial radius `a` and flattening `f`.
     The transformation function takes in the planetocentric
     rectangular coordinates `px`, `py`, `pz` and returns the planetodetic `longitude`, `latitude`,
     and `altitude`. The units of `a`, `px`, `py`, `pz`, and `altitude` are all the same and
@@ -26,14 +26,14 @@ def get_nonflat_pc2pd(a, f):
     Parameters
     ----------
     a : float
-        equitorial radius
+        equatorial radius
     f : float
         flattening parameter
     
     Returns
     -------
     pcf2pd : callable
-        The planetocentric to planetodetic coordinate trnasformation function.
+        The planetocentric to planetodetic coordinate transformation function.
     """
     def pcf2pd(px, py, pz):
         return np.array(erfa.gc2gde(a, f, np.array([px, py, pz])))
@@ -45,12 +45,12 @@ def get_flat_pc2pd():
     flat planetary model.
     The transformation function takes in the planetocentric
     rectangular coordinates `px`, `py`, `pz` and returns a constant 0 for `longitude`, 
-    -pi/2 for `latitutde`, and uses `pz` for `altitude`.
+    -pi/2 for `latitude`, and uses `pz` for `altitude`.
  
     Returns
     -------
     pcf2pd : callable
-        The planetocentric to planetodetic coordinate trnasformation function.
+        The planetocentric to planetodetic coordinate transformation function.
     """
     def pcf2pd(px, py, pz):
         return np.array([0.0, -np.pi/2, pz])
@@ -59,7 +59,7 @@ def get_flat_pc2pd():
 def get_nonflat_pd2pc(a, f):
     """
     Generate a planetodetic to planetocentric coordinate transformation function for an
-    ellipsoidal planetary model, parameterized by equitorial radius`$a` and flattening `f`.
+    ellipsoidal planetary model, parameterized by equatorial radius`$a` and flattening `f`.
     The transformation function takes in the position defined by the planetodetic 
     `longitude`, `latitude`, and `altitude` and returns the planetocentric
     rectangular coordinates `px`, `py`, `pz`.
@@ -70,14 +70,14 @@ def get_nonflat_pd2pc(a, f):
     Parameters
     ----------
     a : float
-        equitorial radius
+        equatorial radius
     f : float
         flattening parameter
     
     Returns
     -------
     pcf2pd : callable
-        The planetodetic to planetocentric coordinate trnasformation function.
+        The planetodetic to planetocentric coordinate transformation function.
     """
     def pd2pcf(longitude, latitude, altitude):
         return np.array(erfa.gd2gce(a, f, longitude, latitude, altitude))
@@ -89,12 +89,12 @@ def get_flat_pd2pc():
     flat planetary model.
     The transformation function takes in the planetocentric
     rectangular coordinates px, py, pz and returns a constant 0 longitude, 
-    -90 degree latitutde, and uses pz for altitude.
+    -90 degree latitude, and uses pz for altitude.
     
     Returns
     -------
     pcf2pd : callable
-        The planetocentric to planetodetic coordinate trnasformation function.
+        The planetocentric to planetodetic coordinate transformation function.
     """
     def pd2pcf(longitude, latitude, altitude):
         return np.array([longitude, latitude, altitude])
@@ -110,7 +110,7 @@ class Planetodetic:
         Parameters
         ----------
         a : float
-            equitorial radius
+            equatorial radius
         f : float
             flattening parameter
[…]
