### 2022-02-25 09ddce0
Clean up doc string formatting
