+++ simupy_flight/__init__.py
@@ -15,12 +15,15 @@ earth_mean_radius = 6371007.1809
 # kinematic functions
 def get_nonflat_pc2pd(a, f):
     """
-    Generate a planetocentric to planetodetic coordinate transformation function for an
-    ellipsoidal planetary model, parameterized by equatorial radius `a` and flattening `f`.
-    The transformation function takes in the planetocentric
-    rectangular coordinates `px`, `py`, `pz` and returns the planetodetic `longitude`, `latitude`,
-    and `altitude`. The units of `a`, `px`, `py`, `pz`, and `altitude` are all the same and
-    assumed to be meters to be consistent with other kinematics.
+    Generate a planetocentric-to-planetodetic coordinate transformation function for an
+    ellipsoidal planetary model, parameterized by equatorial radius `a` and flattening
+    `f`.
+
+    The transformation function takes in the planetocentric rectangular coordinates
+    `px`, `py`, `pz` and returns the planetodetic `longitude`, `latitude`,
+    and `altitude`. The units of `a`, `px`, `py`, `pz`, and `altitude` are all the same
+    and assumed to be meters to be consistent with other kinematics.
+
     Uses the ERFA implementation of Fukushima's method.
 
     Parameters
@@ -29,11 +32,11 @@ def get_nonflat_pc2pd(a, f):
         equatorial radius
     f : float
         flattening parameter
-    
+
     Returns
     -------
     pcf2pd : callable
-        The planetocentric to planetodetic coordinate transformation function.
+        The planetocentric-to-planetodetic coordinate transformation function.
     """
     def pcf2pd(px, py, pz):
         return np.array(erfa.gc2gde(a, f, np.array([px, py, pz])))
@@ -41,16 +44,17 @@ def get_nonflat_pc2pd(a, f):
 
 def get_flat_pc2pd():
     """
-    Generate a planetocentric to planetodetic coordinate transformation function for a
+    Generate a planetocentric-to-planetodetic coordinate transformation function for a
     flat planetary model.
-    The transformation function takes in the planetocentric
-    rectangular coordinates `px`, `py`, `pz` and returns a constant 0 for `longitude`, 
-    -pi/2 for `latitude`, and uses `pz` for `altitude`.
- 
+
+    The transformation function takes in the planetocentric rectangular coordinates
+    `px`, `py`, `pz` and returns a constant 0 for `longitude`, -pi/2 for `latitude`, and
+    uses `pz` for `altitude`.
+
     Returns
     -------
     pcf2pd : callable
-        The planetocentric to planetodetic coordinate transformation function.
+        The planetocentric-to-planetodetic coordinate transformation function.
     """
     def pcf2pd(px, py, pz):
         return np.array([0.0, -np.pi/2, pz])
@@ -58,13 +62,15 @@ def get_flat_pc2pd():
 
 def get_nonflat_pd2pc(a, f):
     """
-    Generate a planetodetic to planetocentric coordinate transformation function for an
-    ellipsoidal planetary model, parameterized by equatorial radius`$a` and flattening `f`.
-    The transformation function takes in the position defined by the planetodetic 
-    `longitude`, `latitude`, and `altitude` and returns the planetocentric
-    rectangular coordinates `px`, `py`, `pz`.
-    The units of `a`, `px`, `py`, `pz`, and `altitude` are all the same and
-    assumed to be meters to be consistent with other kinematics.
+    Generate a planetodetic-to-planetocentric coordinate transformation function for an
+    ellipsoidal planetary model, parameterized by equatorial radius `a` and flattening
+    `f`.
+
+    The transformation function takes in the position defined by the planetodetic
+    `longitude`, `latitude`, and `altitude` and returns the planetocentric rectangular
[…]
