### 2022-02-26 e12573f
F16 model uses good combination of simupy_flight API and directly applying dave-ml code
