### 2022-02-26 7118959
route the aero controlled inputs to tot_aero_forces_moments rather than just on the input aero
