+++ nesc_test_cases/F16_model.py
@@ -24,18 +24,28 @@ from nesc_testcase_helper import ft_per_m, kg_per_slug
 
 N_per_lbf = 4.44822
 
+# Use the F16_aero parameter data from DaveML-generated file
+# convert from english units to SI
 S_A = F16_aero.sref/(ft_per_m**2)
 b_l = F16_aero.bspan/ft_per_m
 c_l = F16_aero.cbar/ft_per_m
 a_l = b_l
 
 class F16(simupy_flight.Vehicle):
-    def __init__(self, CG_PCT_MAC=25):
+    def __init__(self, CG_PCT_MAC=25, use_reported_value=True):
+        # Use the F16_inertia parameter data from DaveML-generated file
+        # convert from english units to SI
         inertia_output = F16_inertia.F16_inertia(CG_PCT_MAC)
         Ixx, Iyy, Izz, Izx, Ixy, Iyz = inertia_output[:6]*kg_per_slug/(ft_per_m**2)
-        #m = inertia_output[6]*kg_per_slug
-        m = 637.26*kg_per_slug #slug
         y_com, z_com, x_com = inertia_output[7:]/ft_per_m
+        # the mass value from the DaveML file does not match the reported value;
+        # we use the reported value by default
+        if use_reported_value:
+            m = 637.26*kg_per_slug #slug
+        else:
+            m = inertia_output[6]*kg_per_slug
+
+        # initialize default vehicle data, including pointing to the propulsion model
         super().__init__(
             m=m,
             I_xx=Ixx, I_yy=Iyy, I_zz=Izz,
@@ -54,9 +64,9 @@ class F16(simupy_flight.Vehicle):
         )
 
     def tot_aero_forces_moments(self, qbar, Ma, Re, V_T, alpha, beta, p_B, q_B, r_B, el,
-                                ail, rdr,):# *args):
-        #print("hello")
-        # el, ail, rdr = args[0][:-1]
+                                ail, rdr,):
+        # use F16_aero model from DaveML-generated file. Inputs converted to english
+        # units. Control surfaces input signals are routed and used here.
         cbar, bspan, sref, cx, cy, cz, cl, cm, cn = F16_aero.F16_aero(V_T*ft_per_m, alpha*180/np.pi, beta*180/np.pi, p_B, q_B, r_B, el, ail, rdr)
         forces_moments = np.empty(6)
         body_force_coeffs = np.array([cx, cy, cz])
@@ -68,6 +78,8 @@ class F16(simupy_flight.Vehicle):
         return forces_moments
 
     def prop_model(self, *args):
+        # use F16_prop model from DaveML-generated file.
+
         throttle = args[-1]
         alt = args[simupy_flight.Vehicle.h_D_arg_idx]*ft_per_m
         Ma = args[simupy_flight.Vehicle.Ma_arg_idx]
