+++ nesc_test_cases/nesc_case01.py
@@ -1,9 +1,29 @@
+"""
+===================================================
+Case 1: Dropped sphere with no drag
+===================================================
+
+==============  ===============
+Verifies        Gravitation, translational EOM
+Gravitation     J2
+Geodesy         WGS-84 rotating
+Atmosphere      US 1976 STD
+Winds           still air
+Vehicle         Dragless spheree
+Notes           Drag coefficient set to zero
+==============  ===============
+"""
+
 from simupy.block_diagram import BlockDiagram
 import simupy_flight
 import numpy as np
 
 from nesc_testcase_helper import plot_nesc_comparisons, int_opts, benchmark
 from nesc_testcase_helper import ft_per_m
+
+# %%
+# Construct the Planet model according to the
+# configuration listed above
 
 planet = simupy_flight.Planet(
     gravity=simupy_flight.earth_J2_gravity,
@@ -16,7 +36,17 @@ planet = simupy_flight.Planet(
     )
 )
 
+# %%
+# Since there is no dynamics model (aerodynamic drag is set to zero)
+# there is no need to define a Vehicle model. So we construct the
+# BlockDiagram with just the planet object
+
 BD = BlockDiagram(planet)
+
+# %%
+# List the planetodetic initial conditions and use the
+# `ic_from_planetodetic` method to transform it to the inertial
+# coordinate system needed for the initial condition
 
 lat_ic = 0.*np.pi/180
 long_ic = 0.*np.pi/180
@@ -33,6 +63,9 @@ omega_Z_ic = 0.
 
 planet.initial_condition = planet.ic_from_planetodetic(long_ic, lat_ic, h_ic, V_N_ic, V_E_ic, V_D_ic, psi_ic, theta_ic, phi_ic)
 planet.initial_condition[-3:] = omega_X_ic, omega_Y_ic, omega_Z_ic
+
+# %%
+# Simulate and assess the results
 
 with benchmark() as b:
     res = BD.simulate(30, integrator_options=int_opts)
