### 2022-02-27 74fa9cc
annotations on a few, adding header+table to all cases
