+++ nesc_test_cases/nesc_case05.py
@@ -1,3 +1,18 @@
+"""
+=====================================================
+Case 5: Dropped sphere over rotating, spherical Earth
+=====================================================
+
+==============  ===============
+Verifies        Earth rotation
+Gravitation     1/R^2
+Geodesy         Round rotating
+Atmosphere      US 1976 STD
+Winds           still air
+Vehicle         Sphere with constant C_D
+==============  ===============
+"""
+
 from simupy.block_diagram import BlockDiagram
 import simupy_flight
 import numpy as np
