+++ nesc_test_cases/nesc_case13p4.py
@@ -1,3 +1,24 @@
+"""
+=========================================================
+Case 13.4: Lateral offset manuever of a subsonic aircraft
+=========================================================
+
+==============  ===============
+Verifies        Multidimensional table look-up
+Gravitation     J2
+Geodesy         WGS-84 rotating
+Atmosphere      US 1976 STD
+Winds           still air
+Vehicle         F-16 with simple auto-pilot
+Notes           Initially straight & level. t=20s, 2000-ft lateral course offset.
+==============  ===============
+
+For the manuevering examples, the BlockDiagram from case 11 is modified to replace the
+controller with the auto-pilot configuration and generate the appropriate command
+signals.
+"""
+
+
 from simupy import systems
 import numpy as np
 from scipy import interpolate
