### 2022-02-27 b9bde28
A few minor paper updates
- Create a separate Statement of Need section
- Add a Description section for more detailed implementation info and
  additional features
