+++ simupy_flight/parse_daveml.py
@@ -8,14 +8,20 @@ from simupy import codegen
 
 delimeter_regex = re.compile("\s+|(?<!\d)[.](?!\d)|,")
 
-def isTag(node, tagsuffix=''):
-    return not isinstance(node, etree._Comment) and isinstance(node.tag, str) and node.tag.endswith(tagsuffix)
 
-def filterTags(nodes, tag=''):
+def isTag(node, tagsuffix=""):
+    return (
+        not isinstance(node, etree._Comment)
+        and isinstance(node.tag, str)
+        and node.tag.endswith(tagsuffix)
+    )
+
+
+def filterTags(nodes, tag=""):
     return [node for node in nodes if isTag(node, tag)]
 
 
-Clip = sp.Function('clip')
+Clip = sp.Function("clip")
 
 run_check_code_template = """
 def run_checks():
@@ -24,6 +30,8 @@ def run_checks():
             raise ValueError("Check for {funcname} failed!")
     print("All checks for {funcname} passed.")
 """
+
+
 class ProcessDaveML:
     def __init__(self, filename):
         self.input_vars = []
@@ -36,90 +44,107 @@ class ProcessDaveML:
         self.extra_tables = {}
         self.check_data = []
 
-        with open(filename, 'r') as xmlfile:
-            text_data = xmlfile.read().replace('\n','')
+        with open(filename, "r") as xmlfile:
+            text_data = xmlfile.read().replace("\n", "")
 
         xml_parsed_data = etree.XML(text_data)
         self.process_xml(xml_parsed_data)
 
         path, fname = os.path.split(filename)
-        funcname = fname.split('.')[0]
+        funcname = fname.split(".")[0]
         # mkdir? and add __init__.py
         # no, it could just be one file with everything
         # will need to dump spline data, so could do that as a dir or just as names
         # also need to save constants
-        constant_modules = {constant: '' for constant in self.constants }
-        function_modules = {spline.func: '' for spline in self.splines }
+        constant_modules = {constant: "" for constant in self.constants}
+        function_modules = {spline.func: "" for spline in self.splines}
         MNPP = codegen.ModuleNumPyPrinter(
-            constant_modules = constant_modules,
-            function_modules = function_modules,
+            constant_modules=constant_modules,
+            function_modules=function_modules,
         )
 
         pre_lines = []
         if self.splines:
-            pre_lines.append('import ndsplines')
-        pre_lines.append('')
+            pre_lines.append("import ndsplines")
+        pre_lines.append("")
 
         if self.constants:
-            pre_lines.append('')
+            pre_lines.append("")
             for var, val in self.constants.items():
-                pre_lines.append('%s = %e' % (str(var), val))
-
+                pre_lines.append("%s = %e" % (str(var), val))
 
         array_printer = codegen.ArrayNumPyPrinter()
 
         if self.splines:
-            pre_lines.append('\n')
+            pre_lines.append("\n")
             for var, spline in self.splines.items():
                 base_name = str(var.func)
                 knot_vars = []
                 for knot_idx, knot_arr in enumerate(spline.knots):
-                    knot_var = '%s_knots_%d' % (base_name, knot_idx)
+                    knot_var = "%s_knots_%d" % (base_name, knot_idx)
                     knot_value = array_printer.doprint(knot_arr)
                     knot_vars.append(knot_var)
-                    pre_lines.append('%s = %s' % (knot_var, knot_value))
+                    pre_lines.append("%s = %s" % (knot_var, knot_value))
 
-                knots_to_init_str = '[%s]' % ','.join([knot_var for knot_var in knot_vars])
+                knots_to_init_str = "[%s]" % ",".join(
[…]
