+++ nesc_test_cases/nesc_case13p1.py
@@ -25,15 +25,12 @@ from nesc_testcase_helper import plot_nesc_comparisons, plot_F16_controls, bench
 from nesc_testcase_helper import ft_per_m
 from nesc_case11 import (
     int_opts,
-    get_controller_function,
+    F16ControllerBlock,
     BD,
     spec_ic_args,
     opt_ctrl,
     dim_feedback,
 )
-
-int_opts["max_step"] = 0.0  # 2**-4
-int_opts["name"] = "dop853"
 
 altCmdBlock = systems.SystemFromCallable(
     interpolate.make_interp_spline(
@@ -46,9 +43,7 @@ altCmdBlock = systems.SystemFromCallable(
 )
 
 BD.systems[-3] = altCmdBlock
-BD.systems[2] = systems.SystemFromCallable(
-    get_controller_function(*opt_ctrl, sasOn=True, apOn=True), dim_feedback + 4, 4
-)
+BD.systems[2] = F16ControllerBlock(*opt_ctrl, sasOn=True, apOn=True, event_t=5.)
 
 with benchmark() as b:
     res = BD.simulate(20, integrator_options=int_opts)
