+++ nesc_test_cases/nesc_case13p4.py
@@ -26,7 +26,7 @@ from scipy import interpolate
 from nesc_testcase_helper import plot_nesc_comparisons, plot_F16_controls, benchmark
 from nesc_case11 import (
     int_opts,
-    get_controller_function,
+    F16ControllerBlock,
     BD,
     spec_ic_args,
     opt_ctrl,
@@ -55,9 +55,7 @@ latOffsetBlock = systems.DynamicalSystem(
 )
 
 BD.systems[-2] = latOffsetBlock
-BD.systems[2] = systems.SystemFromCallable(
-    get_controller_function(*opt_ctrl, sasOn=True, apOn=True), dim_feedback + 4, 4
-)
+BD.systems[2] = F16ControllerBlock(*opt_ctrl, sasOn=True, apOn=True, event_t=20.)
 BD.connect(baseChiCmdBlock, latOffsetBlock, inputs=[0])
 BD.connect(
     earth, latOffsetBlock, outputs=[earth.V_N_idx, earth.V_E_idx], inputs=[1, 2]
