### 2022-03-03 38a1f5d
tweaking tolerances for F16 maneuvering cases
