### 2022-03-05 87adff1
Auto-write failing test outputs to regression_data folder
