+++ nesc_test_cases/run_nesc_cases.py
@@ -18,7 +18,7 @@ cases_to_run += ["13p%01d" % num for num in range(1, 5)]
 cases_to_run += ["%02d" % num for num in range(15, 17)]
 
 for case_id in cases_to_run:
-    print("case number:", case_id)
+    print("case number:", case_id, flush=True)
     call(["python", "nesc_case%s.py" % case_id] + sys.argv[1:])
 
-    print("\n\n")
+    print("\n\n", flush=True)
