### 2022-03-10 25a6f85
Adjust test tolerances
- Exclude 15/16 from automatic tests for now
- Test only Vehicle/Planet output cols for case 13*
