### 2022-05-04 51dfaf4
Set up sphinx docs with gallery
