### 2022-05-04 2d69542
A few minor docs improvements
- Add links to models implemented in the library (gravity, aero, etc.)
- Add a `# %%` in a few key places so plots always follow code in the
  gallery
- Use math for rendering some things
