+++ nesc_test_cases/nesc_case07.py
@@ -9,7 +9,7 @@ Gravitation     J2
 Geodesy         WGS-84 rotating
 Atmosphere      US 1976 STD
 Winds           steady 20 ft/s from due west
-Vehicle         Sphere with constant C_D
+Vehicle         Sphere with constant :math:`C_D`
 ==============  ===============
 """
 
@@ -93,5 +93,7 @@ planet.initial_condition[-3:] = omega_X_ic, omega_Y_ic, omega_Z_ic
 
 with benchmark() as b:
     res = BD.simulate(30, integrator_options=int_opts)
+
+# %%
 
 plot_nesc_comparisons(res, "07")
