+++ .github/workflows/docs.yml
@@ -1,0 +1,36 @@
+name: Docs build
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+
+    - name: Set up Python
+      uses: actions/setup-python@v3
+      with:
+        python-version: '3.9'
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install .[test,docs]
+
+    - name: Build documentation
+      run: |
+        cd docs
+        make html
+
+    - name: Deploy to gh pages
+      uses: JamesIves/github-pages-deploy-action@v4.3.3
+      if: ${{ github.ref == 'refs/heads/master' }}
+      with:
+        branch: gh-pages
+        folder: ./docs/_build/html
