### 2022-05-04 af1422a
Switch to Python 3.8 for docs build
