### 2022-05-04 abb5ca0
Try make clean before docs build
