+++ .github/workflows/docs.yml
@@ -26,6 +26,7 @@ jobs:
     - name: Build documentation
       run: |
         cd docs
+        make clean
         make html
 
     - name: Deploy to gh pages
