+++ .github/workflows/docs.yml
@@ -29,9 +29,9 @@ jobs:
         make clean
         make html
 
-    - name: Deploy to gh pages
-      uses: JamesIves/github-pages-deploy-action@v4.3.3
-      if: ${{ github.ref == 'refs/heads/master' }}
-      with:
-        branch: gh-pages
-        folder: ./docs/_build/html
+    # - name: Deploy to gh pages
+    #   uses: JamesIves/github-pages-deploy-action@v4.3.3
+    #   if: ${{ github.ref == 'refs/heads/master' }}
+    #   with:
+    #     branch: gh-pages
+    #     folder: ./docs/_build/html
