### 2022-05-04 a510f51
dbg: list files in examples dir
