### 2022-05-04 caa0192
Fix help string for no-test flag
