### 2022-05-04 7eec3f7
Fix ignore_pattern
Pattern is matched against full path - in github actions the cwd is
/home/runner/work/... so the `(run)` was matching everything
