+++ .github/workflows/docs.yml
@@ -26,15 +26,11 @@ jobs:
     - name: Build documentation
       run: |
         cd docs
-        make clean
-        pwd
-        ls -lAh .
-        ls -lAh ../nesc_test_cases
         make html
 
-    # - name: Deploy to gh pages
-    #   uses: JamesIves/github-pages-deploy-action@v4.3.3
-    #   if: ${{ github.ref == 'refs/heads/master' }}
-    #   with:
-    #     branch: gh-pages
-    #     folder: ./docs/_build/html
+    - name: Deploy to gh pages
+      uses: JamesIves/github-pages-deploy-action@v4.3.3
+      if: ${{ github.ref == 'refs/heads/master' }}
+      with:
+        branch: gh-pages
+        folder: ./docs/_build/html
