+++ nesc_test_cases/nesc_case15.py
@@ -1,6 +1,6 @@
 """
 ===================================================
-Case 15: Circular F-15 flight around North Pole
+Case 15: Circular F-16 flight around North Pole
 ===================================================
 
 ==============  ===============
