+++ nesc_test_cases/README.rst
@@ -3,3 +3,7 @@ NESC Test Cases
 
 These are implementations of the NESC test cases using the SimuPy Flight Vehicle
 Toolkit.
+
+Additional information about the cases can be found at the NASA Engineering and
+Safety Center's `"Six degree-of-freedom (6-DOF) Flight Simulation Test Cases."
+<https://nescacademy.nasa.gov/flightsim/>`_.
