+++ .github/workflows/docs.yml
@@ -28,9 +28,16 @@ jobs:
         cd docs
         make html
 
+    - name: Upload artifact
+      uses: actions/upload-artifact@v3
+      with:
+        name: docs
+        path: ./docs/_build/
+
     - name: Deploy to gh pages
       uses: JamesIves/github-pages-deploy-action@v4.3.3
       if: ${{ github.ref == 'refs/heads/master' }}
       with:
         branch: gh-pages
         folder: ./docs/_build/html
+
