### 2022-05-09 #4299 @ixjlyons
@athulpg007: thanks for your review.

Your suggestions have been implemented here: https://github.com/nasa/simupy-flight/pull/6

> 1. It turns out all the test cases in this package are already well documented in this [NASA report](https://nescacademy.nasa.gov/src/flightsim/Reports/NASA-TM-2015-218675-EOM_checkcase_appendices.pdf). I suggest adding this to "NESC Test Cases" section to the README file, list the titles of all 16 test cases used, and providing a link to this report. This will greatly help the reader understand the kind of problems that can be solved using this package and provide them with a well-documented list of use-cases both in the package and the NASA report. Optionally, you can also include the report PDF in the repository so users have it offline when they clone the repo.

We had been thinking about making a documentation page to enumerate the examples and show the output, so we've put that here: https://nasa.github.io/simupy-flight/nesc_test_cases/index.html

A link to the Appendix of the Tech Memo describing the test cases was added to the README as well.

> 2. I suggest adding an "Example" section in the paper with the following simple example.
> [snip]

Our understanding of the JOSS format is that the paper is intended to be somewhat of a minimal citable description of the software and motivation for writing it (see the last paragraph of ["What should my paper contain"](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)) -- sample code and outputs seems to be more in the realm of documentation or a full software paper. Does having the example gallery in the documentation help address your concern? @prashjha any thoughts here?

> 3. I ran almost all the test cases supplied, except some which required more time than a few minutes. All the cases I ran exited normally. It appears that `matplotlib`, `pandas`, and `ndsplines` are needed for these examples. Consider adding them to the list of `install_requires` in setup.py.

We did intentionally include only strict dependencies for the library in `install_requires` in case a user doesn't need the others. There is however an "extra" for running the test cases (e.g. `pip install .[test]`). This was probably not that clear in the README previously, so instructions for installing those dependencies has been moved to be just before instructions for running the test cases.

> 4. In readme, the [SimuPy's DynamicalSystem interface](https://simupy.readthedocs.io/en/latest/api/api.html) link is broken. (extra '>'). Please correct this.

Fixed, thanks.

> 5. Please consider adding an API reference for this package using sphinx or something similar. It would be nice to have a dedicated readthedocs for this package with the API reference. If not, the documentation from sphinx can be included it in a docs folder in the repository.

[Done](https://nasa.github.io/simupy-flight/api.html). Like I said, we had been thinking about doing this, and your review was helpful for motivation.

> 6. I would like to know if there are any other tools out there which you are aware of that offer similar functionality. If so, how is this work different compared to existing tools?

The other tools used for the NESC report are described in [Appendix B.6](https://ntrs.nasa.gov/citations/20150001264). The only one that is available publicly is JSBSim. SimuPy Flight is somewhat leaner and takes advantage of existing libraries for ODE integration, atmospherics, etc.
