+++ _downloads/20d15bc8432015220932315296f25c8b/nesc_case16.py
@@ -1,6 +1,6 @@
 """
 =================================================================================
-Case 16: Circular F-15 flight around Equator/dateline intersection
+Case 16: Circular F-16 flight around Equator/dateline intersection
 =================================================================================
 
 ==============  ===============
