### 2022-05-10 #4299 @aliaksei135
Hello @ixjlyons, I have gone through the bulk of the testing code.

## Testing
I have run the test cases and have had a few failures and things to note during the testing:
- cases 5, 7, 8, 13p1 have failed
- cases 15 and 16 do not indicate whether they have passed or not
- I can see that the test case runs in the documentation have the same errors, but an explicit note that the "missing_*" and `RuntimeWarning` are expected would be good. I assume the `RuntimeWarning`s are caused by zero drag being enforced.
- Is there any appetite to automate the rest of test runs? Not a blocker for the review, but would make sense in the long term. Seems it is only case 1 that is being run in Github Actions.

<summary>Full output</summary>

```
(sf_review) P:\Dev\PycharmProjects\simupy-flight>python nesc_test_cases/run_nesc_cases.py

case number: 01
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: invalid value encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
time to simulate: 1.852 s
Regression test PASSED
missing eulerAngle_deg_Yaw for SIM 03
missing eulerAngle_deg_Pitch for SIM 03
missing eulerAngle_deg_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Pitch for SIM 03
missing bodyAngularRateWrtEi_deg_s_Yaw for SIM 03
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_X for SIM 03
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Y for SIM 03
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02
missing eiPosition_ft_Z for SIM 03

case number: 02
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: invalid value encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\dynamics.py:22: RuntimeWarning: divide by zero encountered in double_scalars
  x15 = numpy.select([numpy.greater(V_T, 0.0)], [(1/2)/V_T], default=0.0)
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: divide by zero encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
time to simulate: 2.688 s
Regression test PASSED
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02

case number: 03
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: invalid value encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\dynamics.py:22: RuntimeWarning: divide by zero encountered in double_scalars
  x15 = numpy.select([numpy.greater(V_T, 0.0)], [(1/2)/V_T], default=0.0)
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: divide by zero encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
time to simulate: 2.689 s
Regression test PASSED
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02

case number: 04
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: invalid value encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\dynamics.py:22: RuntimeWarning: divide by zero encountered in double_scalars
  x15 = numpy.select([numpy.greater(V_T, 0.0)], [(1/2)/V_T], default=0.0)
time to simulate: 2.753 s
Regression test PASSED
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 05
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: invalid value encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\dynamics.py:22: RuntimeWarning: divide by zero encountered in double_scalars
  x15 = numpy.select([numpy.greater(V_T, 0.0)], [(1/2)/V_T], default=0.0)
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: divide by zero encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
time to simulate: 2.723 s
Regression test FAILED
col 24 failed
Writing output data for comparison
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 06
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\kinematics.py:198: RuntimeWarning: invalid value encountered in double_scalars
  x60 = x6*(2*x26*x54 + x58*(-x13 + x14) + x59*(x33 + x55))/x53
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\dynamics.py:22: RuntimeWarning: divide by zero encountered in double_scalars
  x15 = numpy.select([numpy.greater(V_T, 0.0)], [(1/2)/V_T], default=0.0)
time to simulate: 2.706 s
Regression test PASSED
missing eulerAngle_deg_Yaw for SIM 03
missing eulerAngle_deg_Pitch for SIM 03
missing eulerAngle_deg_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Pitch for SIM 03
missing bodyAngularRateWrtEi_deg_s_Yaw for SIM 03
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_X for SIM 03
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Y for SIM 03
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02
missing eiPosition_ft_Z for SIM 03

case number: 07
time to simulate: 2.648 s
Regression test FAILED
col 23 failed
col 24 failed
col 34 failed
Writing output data for comparison
missing eulerAngle_deg_Yaw for SIM 03
missing eulerAngle_deg_Pitch for SIM 03
missing eulerAngle_deg_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Pitch for SIM 03
missing bodyAngularRateWrtEi_deg_s_Yaw for SIM 03
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_X for SIM 03
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Y for SIM 03
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02
missing eiPosition_ft_Z for SIM 03

case number: 08
time to simulate: 2.547 s
Regression test FAILED
col 23 failed
col 24 failed
Writing output data for comparison
missing eulerAngle_deg_Yaw for SIM 03
missing eulerAngle_deg_Pitch for SIM 03
missing eulerAngle_deg_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Pitch for SIM 03
missing bodyAngularRateWrtEi_deg_s_Yaw for SIM 03
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_X for SIM 03
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Y for SIM 03
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02
missing eiPosition_ft_Z for SIM 03

case number: 09
time to simulate: 2.657 s
Regression test PASSED
missing eulerAngle_deg_Yaw for SIM 03
missing eulerAngle_deg_Pitch for SIM 03
missing eulerAngle_deg_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Pitch for SIM 03
missing bodyAngularRateWrtEi_deg_s_Yaw for SIM 03
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_X for SIM 03
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Y for SIM 03
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02
missing eiPosition_ft_Z for SIM 03

case number: 10
time to simulate: 2.642 s
Regression test PASSED
missing eulerAngle_deg_Yaw for SIM 03
missing eulerAngle_deg_Pitch for SIM 03
missing eulerAngle_deg_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Roll for SIM 03
missing bodyAngularRateWrtEi_deg_s_Pitch for SIM 03
missing bodyAngularRateWrtEi_deg_s_Yaw for SIM 03
missing eiPosition_ft_X for SIM 01
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_X for SIM 03
missing eiPosition_ft_Y for SIM 01
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Y for SIM 03
missing eiPosition_ft_Z for SIM 01
missing eiPosition_ft_Z for SIM 02
missing eiPosition_ft_Z for SIM 03

case number: 11
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
time to simulate: 69.494 s
Regression test FAILED
col 30 failed
Writing output data for comparison
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 13p1
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
time to simulate: 28.837 s
Regression test FAILED
col 34 failed
col 36 failed
col 38 failed
Writing output data for comparison
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 13p2
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
time to simulate: 28.910 s
Regression test PASSED
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 13p3
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
time to simulate: 45.113 s
Regression test PASSED
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 13p4
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
time to simulate: 87.702 s
Regression test PASSED
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 15
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
Optimization terminated successfully.
         Current function value: 5.307430
         Iterations: 270
         Function evaluations: 553
pitch: 2.6742e+00  roll: 0.0000e+00  longStk: 13.0082  throttle: 13.8357
accelerations:
 [[-5.30742973e+00 -1.50155598e-07  5.68309078e-09]
 [-6.38762624e-07 -6.51306706e-08  5.56268870e-09]]
time to simulate: 713.195 s
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02

case number: 16
Optimization terminated successfully.
         Current function value: 0.006500
         Iterations: 281
         Function evaluations: 551
pitch: 2.6351e+00  roll: 0.0000e+00  longStk: 12.9236  throttle: 13.7561
accelerations:
 [[-4.59610657e-03  4.59609475e-03  2.29389308e-09]
 [-2.98437294e-09  2.76702995e-09  2.12618593e-09]]
Optimization terminated successfully.
         Current function value: 0.000000
         Iterations: 281
         Function evaluations: 522
pitch: 2.6659e+00  roll: 0.0000e+00  longStk: 12.9899  throttle: 13.8207
accelerations:
 [[-7.73936470e-12  3.10714995e-12 -3.37891466e-12]
 [ 1.83141906e-11 -1.04603109e-12  1.45029217e-09]]
time to simulate: 387.504 s
missing eiPosition_ft_X for SIM 02
missing eiPosition_ft_Y for SIM 02
missing eiPosition_ft_Z for SIM 02
```

---

## Parsing DaveML

- The packages `python-libsbml` and `lxml` are required, but do not seem to be installed anywhere
- In the `derivation` extras package, `jupyter-lab` is not a valid package. You probably meant just  `jupyter`
- In `process_NESC_DaveML.py`, you should use the fully qualified package name to import the `parse_daveml`, otherwise it only runs from a single dir:
```python
from parse_daveml import ProcessDaveML
# Should be:
from simupy_flight.parse_daveml import ProcessDaveML
```
- The paths in `process_NESC_DaveML.py` seem to have duplicated the first directory in all paths. I cannot find where these paths would correspond to and suspect this is a mistake.
- Running `process_NESC_DaveML.py` results in an error within the  `simupy` code generation module:
<summary>Error Traceback</summary>

```
(sf_review) P:\Dev\PycharmProjects\simupy-flight\NESC_data\All_models>python ..\..\nesc_test_cases\process_NESC_DaveML.py
Skipping header
P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\parse_daveml.py:255: UserWarning: trimmedTheta variable already in symbol table!
  warnings.warn(
Completed
Traceback (most recent call last):
  File "..\..\nesc_test_cases\process_NESC_DaveML.py", line 23, in <module>
    ProcessDaveML(filename)
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy_flight\parse_daveml.py", line 110, in __init__
    code = printer.codeprint(
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy\codegen.py", line 306, in codeprint
    lines.append(self.doprint(*func_arg_expr))
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy\codegen.py", line 377, in doprint
    funcbody.append(self._exprrepr(Assignment(lhs, rhs)))
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\sympy\printing\codeprinter.py", line 100, in doprint
    lines = self._print(expr).splitlines()
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\sympy\printing\printer.py", line 287, in _print
    return getattr(self, printmethod)(expr, **kwargs)
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy\codegen.py", line 271, in _print_Assignment
    return super()._print_Assignment(expr)
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy\codegen.py", line 201, in _print_Assignment
    return super()._print_Assignment(expr)
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\sympy\printing\codeprinter.py", line 327, in _print_Assignment
    rhs_code = self._print(rhs)
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\sympy\printing\printer.py", line 287, in _print
    return getattr(self, printmethod)(expr, **kwargs)
  File "P:\Dev\PycharmProjects\sf_review\lib\site-packages\simupy\codegen.py", line 256, in _print_Function
    if module != '':
UnboundLocalError: local variable 'module' referenced before assignment
```

You should consider adding `daveml` as an additional extras package with the dependencies above. There also doesn't seem to be explicit instructions for testing of the DaveML module.

Let me know if you need more input from my side or have any questions.
