### 2022-05-12 #4299 @athulpg007
@ixjlyons Thanks for incorporating my suggestions. The examples [https://nasa.github.io/simupy-flight/nesc_test_cases/index.html](https://nasa.github.io/simupy-flight/nesc_test_cases/index.html) and the API section is excellent and very useful.

I recommend that you include an "Examples" section in the paper. It can just contain one line with a link to the examples in the documentation.

I have concluded my review and am happy to accept this submission.
