### 2022-05-13 0773675
Add test result summary. Enable all tests in gh actions
