+++ .github/workflows/test.yml
@@ -1,4 +1,4 @@
-name: Test
+name: Regression test
 
 on:
   push:
@@ -8,12 +8,13 @@ on:
 
 jobs:
   build:
-
-    runs-on: ubuntu-latest
+    name: Regression test
+    runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
         python-version: ["3.8", "3.9", "3.10"]
+        os: [ubuntu-latest, windows-latest, macos-latest]
 
     steps:
     - uses: actions/checkout@v3
@@ -30,4 +31,4 @@ jobs:
 
     - name: Run tests
       run: |
-        python nesc_test_cases/nesc_case01.py
+        python nesc_test_cases/run_nesc_cases.py
