### 2022-05-14 18eeecc
Generate unique artifacts for each job
