### 2022-05-17 ffa7521
Fix extras_require for derivation. Remove stale __init__.py
