### 2022-05-20 59b1fc5
write norms to file
