### 2022-05-23 3c8d06d
more norms, traj res
