### 2022-06-10 26fe0d1
Add run_checks to F16 model modules
