+++ nesc_test_cases/process_NESC_DaveML.py
@@ -2,8 +2,10 @@ import os
 import pathlib
 from simupy_flight.parse_daveml import ProcessDaveML
 
-base_dir = os.path.join(pathlib.Path(__file__).parent.resolve(),
-                        '..', 'NESC_data', 'All_models')
+here = os.path.abspath(os.path.dirname(__file__))
+
+base_dir = os.path.join(here, "..", "NESC_data", "All_models")
+out_dir = here
 
 filenames = [
     os.path.join("F16_package", "F16_S119_source", "F16_control.dml"),
@@ -11,14 +13,7 @@ filenames = [
     os.path.join("F16_package", "F16_S119_source", "F16_prop.dml"),
     os.path.join("F16_package", "F16_S119_source", "F16_gnc.dml"),
     os.path.join("F16_package", "F16_S119_source", "F16_inertia.dml"),
-    os.path.join("two-stage_package", "twostage_aero.dml"),
-    os.path.join("two-stage_package", "twostage_inertia.dml"),
-    os.path.join("two-stage_package", "twostage_prop.dml"),
-    "brick_aero.dml",
-    "brick_inertia.dml",
-    "cannonball_aero.dml",
-    "cannonball_inertia.dml",
 ]
 
 for filename in filenames:
-    ProcessDaveML(os.path.join(base_dir, filename))
+    ProcessDaveML(os.path.join(base_dir, filename), outdir=out_dir)
