### 2022-06-10 d5bd607
Add note to mention runtimewarnings for divide by zero
