### 2022-06-11 ef1945f
Add instructions for checking F16 models
