+++ setup.py
@@ -24,7 +24,7 @@ setup(
     install_requires=[
         "numpy>=1.11.3",
         "scipy>=0.18.1",
-        "simupy>=1.1.1",
+        "simupy>=1.1.2",
         "pyerfa",
         "fluids",
     ],
