### 2022-06-11 #4299 @ixjlyons
Hi @aliaksei135, thanks for the review. Your comments are addressed below.

> I have run the test cases and have had a few failures and things to note during the testing:
> - cases 5, 7, 8, 13p1 have failed
> - cases 15 and 16 do not indicate whether they have passed or not
> - I can see that the test case runs in the documentation have the same errors, but an explicit note that the "missing_*" and `RuntimeWarning` are expected would be good. I assume the `RuntimeWarning`s are caused by zero drag being enforced.
> - Is there any appetite to automate the rest of test runs? Not a blocker for the review, but would make sense in the long term. Seems it is only case 1 that is being run in Github Actions.

We now test only the position and orientation states to tighter tolerances since these "lowest order" states should capture the effect of every computation in the model while integrating out minor numerical differences. We've updated the continuous integration testing to run all cases (including 15 and 16) across a range of platforms and Python versions.

We also added a note to the [README](https://github.com/nasa/simupy-flight#nesc-test-cases) and case 1 to describe why `RuntimeWarning` is being emitted from the generated code.

> - The packages `python-libsbml` and `lxml` are required, but do not seem to be installed anywhere
> - In the `derivation` extras package, `jupyter-lab` is not a valid package. You probably meant just  `jupyter`
>
> You should consider adding `daveml` as an additional extras package with the dependencies above. There also doesn't seem to be explicit instructions for testing of the DaveML module.

Thanks for catching these issues. There's now a `daveml` extras list to install dependencies for the DaveML parsing module. The README has instructions for using it to re-generate the F16 component models, and these generated files now check the model against provided I/O data when they're run.

> - In `process_NESC_DaveML.py`, you should use the fully qualified package name to import the `parse_daveml`, otherwise it only runs from a single dir:
> - The paths in `process_NESC_DaveML.py` seem to have duplicated the first directory in all paths. I cannot find where these paths would correspond to and suspect this is a mistake.
> - Running `process_NESC_DaveML.py` results in an error within the  `simupy` code generation module:

Thanks for catching these as well - they've been fixed. Part of the resolution involved updating simupy, so you'll have to update it to re-run.
