### 2022-07-07 efcb4a1
minor case changes in bib
