+++ paper/references.bib
@@ -40,7 +40,7 @@
 
 @Article{sympy,
   author   = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k, Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K. and Singh, Sartaj and Rathnayake, Thilina and Vig, Sean and Granger, Brian E. and Muller, Richard P. and Bonazzi, Francesco and Gupta, Harsh and Vats, Shivam and Johansson, Fredrik and Pedregosa, Fabian and Curry, Matthew J. and Terrel, Andy R. and Rou\v{c}ka, \v{S}t\v{e}p\'{a}n and Saboo, Ashutosh and Fernando, Isuru and Kulal, Sumith and Cimrman, Robert and Scopatz, Anthony},
-  title    = {SymPy: symbolic computing in Python},
+  title    = {SymPy: symbolic computing in {P}ython},
   doi      = {10.7717/peerj-cs.103},
   issn     = {2376-5992},
   pages    = {e103},
@@ -93,7 +93,7 @@
 
 @Article{margolis2020fuzzy,
   author    = {Margolis, Benjamin WL and Ayoubi, Mohammad A and Joshi, Sanjay S},
-  title     = {Nonlinear model predictive control of reentry vehicles based on Takagi-Sugeno fuzzy models},
+  title     = {Nonlinear model predictive control of reentry vehicles based on {T}akagi-{S}ugeno fuzzy models},
   doi       = {10.1007/s40295-019-00191-2},
   number    = {1},
   pages     = {113--136},
