### 2022-07-07 #4299 @danielskatz
@ixjlyons - I'm the AEiC this week, and working on the processing of your submission.  It all looks good, except I think that there are some small issue in the bib, as indicated in https://github.com/nasa/simupy-flight/pull/12 - If you can merge these (or let me know what you disagree with), we can continue the processing.
