### 2022-07-07 844f8e8
Preserve cases in a couple reference titles
