+++ paper/references.bib
@@ -55,7 +55,7 @@
 
 @software{fluids,
   author={Bell, Caleb},
-  title={fluids: Fluid dynamics component of Chemical Engineering Design Library},
+  title={{fluids}: Fluid dynamics component of {C}hemical {E}ngineering {D}esign {L}ibrary ({ChEDL})},
   year={2016 -- 2021},
   url={https://github.com/CalebBell/fluids},
 }
@@ -133,7 +133,7 @@
 }
 
 @article{ndsplines,
-  title={{ndsplines}: A Python Library for Tensor-Product B-Splines of Arbitrary Dimension},
+  title={{ndsplines}: A {P}ython Library for Tensor-Product {B}-Splines of Arbitrary Dimension},
   author={Margolis, Benjamin WL and Lyons, Kenneth R},
   journal={Journal of Open Source Software},
   volume={4},
